requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.16
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
chromadb>=0.4.18
tiktoken>=0.5.2
openai>=1.3.0
pydantic>=2.5.0
ipython>=8.18.0
